moved constants to `Constants` class
