+++ app/src/main/java/zechs/drive/stream/utils/util/Constants.kt
@@ -1,0 +1,7 @@
+package zechs.drive.stream.utils.util
+
+class Constants {
+    companion object {
+        const val DRIVE_API = "https://www.googleapis.com/drive/v3"
+    }
+}
